requirements.txt
apturl==0.5.2
asn1crypto==0.24.0
astroid==2.3.3
attrs==19.3.0
bcrypt==3.1.7
beautifulsoup4==4.8.2
Brlapi==0.6.6
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
Click==7.0
colorama==0.4.3
colorlog==4.0.2
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
cycler==0.10.0
dataclasses==0.7
decorator==4.4.1
defer==1.0.6
Deprecated==1.2.7
distro-info===0.18ubuntu0.18.04.1
entrypoints==0.3
flake8==3.7.9
flake8-polyfill==1.0.2
future==0.18.2
gitdb2==2.0.6
GitPython==2.1.14
httplib2==0.9.2
idna==2.9
importlib-metadata==1.3.0
ipython-genutils==0.2.0
isort==4.3.21
joblib==0.14.1
jsonschema==3.2.0
jupyter-core==4.6.1
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.1.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
lazy-object-proxy==1.4.3
louis==3.5.0
lxml==4.4.2
macaroonbakery==1.1.3
Mako==1.0.7
mando==0.6.4
MarkupSafe==1.0
matplotlib==3.1.2
mccabe==0.6.1
more-itertools==8.0.2
mysql-connector-python==8.0.18
nbformat==4.4.0
netifaces==0.10.4
numpy==1.18.0
oauth==1.0.1
olefile==0.45.1
pandas==0.25.3
paramiko==2.6.0
pexpect==4.2.1
Pillow==5.1.0
plotly==4.4.1
progress==1.4
protobuf==3.11.1
pycairo==1.16.2
pycodestyle==2.5.0
pycparser==2.19
pycrypto==2.6.1
pycups==1.9.73
pyflakes==2.1.1
PyGithub==1.46
pygobject==3.26.1
PyJWT==1.7.1
pylint==2.4.4
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==2.4.5
pyRFC3339==1.0
pyrsistent==0.15.6
python-apt==1.6.5+ubuntu0.2
python-dateutil==2.8.1
python-debian==0.1.32
pytz==2019.3
pyxdg==0.25
PyYAML==3.12
radon==4.0.0
reportlab==3.4.0
requests==2.23.0
requests-unixsocket==0.1.5
retrying==1.3.3
scikit-learn==0.22
scipy==1.4.1
seaborn==0.9.0
SecretStorage==2.3.1
simplejson==3.13.2
six==1.14.0
smmap2==2.0.5
soupsieve==2.0
ssh-import-id==5.7
sshtunnel==0.1.5
system-service==0.3
systemd-python==234
tabulate==0.8.6
tqdm==4.41.0
traitlets==4.3.3
typed-ast==1.4.1
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.3
wadllib==1.3.2
wily==1.16.0
wrapt==1.12.0
xkit==0.0.0
zipp==0.6.0
zope.interface==4.3.2
